# api/openapi.yaml
openapi: 3.1.0
info:
  title: Trident REST API
  description: |
    Real-time indexing and event querying for Soroban contract events.

    ## Request validation

    Every endpoint validates its parameters before doing any work and answers a
    malformed request with `400` and the `ErrorResponse` envelope, where
    `error.code` is `INVALID_ARGUMENT` and `error.message` names the field that
    failed:

    ```json
    {"error": {"code": "INVALID_ARGUMENT", "message": "limit must be an integer between 1 and 200"}}
    ```

    Rules shared across endpoints:

    - `contractId` — Stellar contract strkey: `C` followed by 55 uppercase
      base32 characters.
    - path ids — canonical UUID v4.
    - ledger ranges — non-negative integers; the upper bound must be greater
      than or equal to the lower bound.
    - `limit` — an integer within the range documented for that endpoint.
    - `cursor` — an opaque cursor from a previous response; anything else is
      rejected rather than silently ignored.
    - `network` — `testnet` or `mainnet`.
    - Unknown query parameters are rejected, so a typo such as `?limitt=5`
      fails loudly instead of silently changing the page size.

    ## Pagination

    Every list endpoint returns the same keyset-pagination envelope: the item
    array, `has_more`, and `next_cursor`.

    `next_cursor` is **always present** in the response object. When there is
    no further page it is the explicit JSON value `null` — it is never omitted:

...

...

...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /v1/contracts/{id}/spec:
    get:
      summary: Get a contract's parsed spec and detected interfaces
      description: Returns the functions captured from a tracked contract's spec (contractmeta / SEP-48) and the standard interfaces detected from them (e.g. SEP-41 token, NFT).
      operationId: getContractSpec
      tags:
        - Contracts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^C[A-Z2-7]{55}$"
          description: Soroban contract address
      responses:
        "200":
          description: >-
            Contract spec and detected interfaces. Cached for 5 minutes
            (issue #221) — see X-Cache.
          headers:
            X-Cache:
              $ref: "#/components/headers/X-Cache"
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractSpecResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  schemas:
...
    ContractSpecResponse:
      type: object
      additionalProperties: false
      required:
        - contract_id
        - network
        - code_hash
        - has_spec
        - contract_type
        - interfaces
        - functions
      properties:
        contract_id:
          type: string
          pattern: "^C[A-Z2-7]{55}$"
          description: Soroban contract address
        network:
          type: string
          enum: [testnet, mainnet]
          description: Network queried
        code_hash:
          type: string
          description: Deployed WASM code hash this spec was parsed from
        has_spec:
          type: boolean
          description: Whether an embedded contractspecv0 section was found
        contract_type:
          type: string
          description: Primary classification derived from detected interfaces (e.g. token, nft, custom)
        interfaces:
          type: array
          items:
            type: string
          description: Every standard interface detected from the contract's spec functions
        functions:
          type: array
          items:
            $ref: "#/components/schemas/ContractSpecFunction"
          description: Functions captured from the contract's spec
    ContractSpecFunction:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          description: Exported function name
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code (e.g., INVALID_ARGUMENT, INTERNAL, UNAVAILABLE, CONFLICT)
            message:
              type: string
              description: Human-readable error message
            request_id:
              type: string
              description: Request ID for debugging
  headers:
    X-RateLimit-Limit:
      description: >-
        Requests allowed per window for this API key's rate-limit tier.
        Present on every response from an endpoint secured by ApiKeyAuth
        (2xx and 429 alike) once a valid X-API-Key was presented.
      required: true
      schema:
        type: integer
        minimum: 0
      example: 50
    X-RateLimit-Remaining:
      description: >-
        Requests remaining in the current window for this API key. 0 on the
        response that triggers a 429.
      required: true
      schema:
        type: integer
        minimum: 0
      example: 12
    X-RateLimit-Reset:
      description: >-
        Unix timestamp (seconds) when the current rate-limit window resets.
      required: true
      schema:
        type: integer
        format: int64
      example: 1732900000
    Retry-After:
      description: >-
        Seconds to wait before retrying. Present on 429 (rate limit
        exceeded, per-API-key or per-IP) and on 503 responses caused by the
        global concurrency cap shedding load. Not present on a 503 caused by
        an unavailable dependency (database/Redis/gRPC backend) — check the
        error envelope's `error.code` to distinguish the two.
      schema:
        type: integer
        minimum: 1
      example: 1
    X-Cache:
      description: >-
        Whether this response was served from the Redis response cache
        (HIT) or freshly computed (MISS). Only emitted by endpoints that
        cache their response.
      required: true
      schema:
        type: string
        enum: [HIT, MISS]
      example: HIT
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimitExceeded:
      description: >-
        Rate limit exceeded for this API key's tier (error.code
        RATE_LIMITED). Carries the same X-RateLimit-* headers as a
        successful response (X-RateLimit-Remaining is 0) plus Retry-After.
      headers:
        X-RateLimit-Limit:
          $ref: "#/components/headers/X-RateLimit-Limit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/X-RateLimit-Remaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/X-RateLimit-Reset"
        Retry-After:
          $ref: "#/components/headers/Retry-After"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServiceUnavailable:
      description: >-
        Service temporarily unavailable — either a dependency (database,
        Redis, gRPC backend) is down, or the server is shedding load under
        the global concurrency cap. When load-shedding is the cause,
        Retry-After is set; otherwise it is absent.
      headers:
        Retry-After:
          $ref: "#/components/headers/Retry-After"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin key for admin endpoints (environment variable)
...